IBM-DS-with-R-2-Data-Analysis-with-R
IBM-DS-with-R-2-Data-Analysis-with-R-Project 
Project introduction: Provide a high level analysis of weather data in JFK Airport. Your stakeholders want to understand the current and historical record of precipitations based on different variables. For now they are mainly interested in a macro-view of JFK Airport Weather, and how it relates to the possibility to rain because it will affect flight delays and etc.